Correctly finds OpenGL ES on RPI even if Open GL is present too.
On Rapsberry Pi, both OpenGL ES and OpenGL can be installed at the same time.
OpenGL libraries can present, even if OpenGL driver is not activated.
This change ensure that the OpenGL ES library is used in priority if present.
I guess that few systems would have both OpenGL and OpenGL ES,
and if OpenGL ES is present that would mean that there is a good reason for it.
So, I think OpenGL ES so be considered first.